Experiment with hippie homegrown web-analytics solution
Analytics are provided by forcing a POST request on each browser visit
to the website, to a dynamic server I co-administrate and which has
the logic to count visits and render ugly graphs:

  http://gasche.ulminfo.fr/neu-prl/statistics/view/

Using javascript and POST means that we probably won't get hits coming
from bots crawling the server, which means (1) good, no over-inflated
numbers but also (2) we probably won't get visits if people read the
posts from Feedly, TT-RSS or other feed aggregators. This should be
taken as incentives to advertise more effectively on social media.

(I could of course use a GET-based solution solution, but I think this
one is conceptually nicer.)