Add fix to make races when deploying much less likely.
Currently, if two commits are merged close together, whichever deploys last wins
this means that if the build on the earlier commit finishes last, it will
overwrite the later commit's build. This makes this much better by checking
right before deploying if our commit is still the latest commit -- if it's not,
we've detected that another build has started (and maybe finished), and we
abort, as the other build should always win.

Note that this still _can_ race -- but the entire other build has to complete
between this new fetch and the push (and the build takes ~2mins). This is in
contrast to the previous behavior where all that needed to happen was the later
commit deployed first.