blog: sampling, add 'uniform random' limitation
Add a tiny note saying the method assumes "programmers will
pick a configuraiton uniformly at random".

We make this assumption because its not obviously wrong and we couldn't
think of anything better. But it might be obviously wrong in some
contexts, and it might be possible to use a stronger assumption in some
other contexts.

Thank you Andrew Cobb