Improve the order of seminars
Upcoming seminars show up first, in chronological order, so the next
seminar is first. Afterwards, past seminars are displayed in reverse
chronological order, so the most recent seminar is first. For example:

 - NEXT SEMINAR
 - NEAR FUTURE
 - FAR FUTURE
 - most recent seminar
 - near past
 - far past

This required a few other changes, such as defining a `seminar` struct
to defer the Scribble rendering until after the partition and sorting.

Note that the sorting is done in the build step, so we also have to
ensure the build system (Travis) builds every night. This also depends
on time zones being properly configured on Travis... (A future
enhancement might be to do the sorting on the client-side in
JavaScript.)

Fixes #167

---

Along the way, I made other minor fixes:

 - fix some CSS
 - fix generated HTML (there was some invalid nesting of p and span)
 - make generated HTML more readable (by inserting new lines)
 - remove some JavaScript from each seminar, which 'collapsed' past
   seminars, because we do that statically now