blog: offsite, reword reason for rejecting Warren Center
Looking back, "price" is not a good reason for rejecting the Warren
Center. The quote we got from Donald said $125 per person for "the day".
Assuming there were no other expenses, that would make the Warren center
a few hundred $$$ cheaper than our MTW offsite (which, after catering,
was about $3800).

Thanks to @aatxe for asking about price (the menu doesn't say) so we
could get the record straight.

What I remember is that we were:

- unhappy with how hard Warren Center was to contact,
- worried about transportation,
- definitely not going to try staying overnight,
- and disappointed with the price quote

We also got this quote before we started looking closer at other venues.

Anyway, the Warren Center is worth reconsidering in the future